Support profiling hook in nds-h
