+++ nds/nds_power.py
@@ -36,8 +36,6 @@ import os
 import re
 import sys
 import time
-import subprocess 
-import shlex
 from collections import OrderedDict
 from pyspark.sql import SparkSession
 from PysparkBenchReport import PysparkBenchReport
@@ -53,48 +51,9 @@ utils_dir = os.path.join(parent_dir, 'utils')
 if utils_dir not in sys.path:
     sys.path.insert(0, utils_dir)
 from spark_utils import setQueryName, clearQueryName
+from profiler import Profiler
 
 check_version()
-
-
-class Profiler:
-    def __init__(self, profiling_hook, output_root):
-        self.profiling_hook = profiling_hook
-        self.output_root = output_root
-        self.is_profiling_enabled = output_root is not None and profiling_hook is not None
-
-        self.query_name = None
-
-    def __call__(self, query_name):
-        self.query_name = query_name
-        return self
-
-    def __enter__(self,):
-        self.start_profiling()
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        self.stop_profiling()
-        self.query_name = None
-
-    def execute_script(self, action):
-        script_path = self.profiling_hook
-        command = f"{script_path} {action} {shlex.quote(self.output_root)} {shlex.quote(self.query_name)}"
-        try:
-            subprocess.run(command, shell=True, check=True)
-        except subprocess.CalledProcessError as e:
-            print(f"Error: Script exited with status {e.returncode}")
-            raise
-
-    def start_profiling(self):
-        if self.is_profiling_enabled:
-            print(f"Profiling started with profiling script: {self.profiling_hook} writing to {self.output_root} for query {self.query_name}.")
-            self.execute_script('start')
-
-    def stop_profiling(self):
-        if self.is_profiling_enabled:
-            self.execute_script('stop')
-            print("Profiling stopped")
 
 
 def split_and_strip(str, delimiter):
